Move JSON schema to "graviton/json-schema".
