+++ app/AppKernel.php
@@ -77,6 +77,7 @@ class AppKernel extends Kernel
             new \Knp\Bundle\GaufretteBundle\KnpGaufretteBundle(),
             new \Eo\AirbrakeBundle\EoAirbrakeBundle(),
             new \HadesArchitect\JsonSchemaBundle\JsonSchemaBundle(),
+            new \Graviton\JsonSchemaBundle\GravitonJsonSchemaBundle(),
         );
 
         if (in_array($this->getEnvironment(), array('dev', 'test', 'oauth_dev'))) {
