--- src/Graviton/GeneratorBundle/Definition/Validator/InvalidJsonException.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * InvalidJsonException class file
- */
-
-namespace Graviton\GeneratorBundle\Definition\Validator;
-
-/**
- * Invalid JSON exception
- *
- * @author   List of contributors <https://github.com/libgraviton/graviton/graphs/contributors>
- * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
- * @link     http://swisscom.ch
- */
-class InvalidJsonException extends \InvalidArgumentException
-{
-}
